+++ src/commands/hunt.ts
@@ -311,7 +311,7 @@ async function run(
     total = 0;
   }
 
-  const earnedXp = await calcEarnedHFMXp(member, total);
+  const earnedXp = await calcEarnedHFMXp(member, total, message.guildId);
 
   if (earnedXp > 0) {
     embed.setFooter({ text: `+${earnedXp.toLocaleString()}xp` });