+++ src/commands/slots.ts
@@ -317,7 +317,8 @@ async function run(
       "slots",
       1,
     );
-    multi = (await getGambleMulti(message.member, message.client as NypsiClient)).multi;
+    multi = (await getGambleMulti(message.member, message.client as NypsiClient, message.guildId))
+      .multi;
     winnings -= bet;
 
     if (multi > 0) {
@@ -396,6 +397,7 @@ async function run(
         message.client as NypsiClient,
         bet,
         multiplier,
+        message.guildId,
       );
 
       if (earnedXp > 0) {